# Faster recurrent policy evaluate actions