--- a/model.py
+++ b/model.py
@@ -82,11 +82,11 @@ class NNBase(nn.Module):
         self._recurrent = recurrent
 
         if recurrent:
-            self.gru = nn.GRUCell(recurrent_input_size, hidden_size)
-            nn.init.orthogonal_(self.gru.weight_ih.data)
-            nn.init.orthogonal_(self.gru.weight_hh.data)
-            self.gru.bias_ih.data.fill_(0)
-            self.gru.bias_hh.data.fill_(0)
+            self.gru = nn.GRU(recurrent_input_size, hidden_size)
+            nn.init.orthogonal_(self.gru.weight_ih_l0.data)
+            nn.init.orthogonal_(self.gru.weight_hh_l0.data)
+            self.gru.bias_ih_l0.data.fill_(0)
+            self.gru.bias_hh_l0.data.fill_(0)
 
     @property
     def is_recurrent(self):
@@ -104,7 +104,9 @@ class NNBase(nn.Module):
 
     def _forward_gru(self, x, hxs, masks):
         if x.size(0) == hxs.size(0):
-            x = hxs = self.gru(x, hxs * masks)
+            x, hxs = self.gru(x.unsqueeze(0), (hxs * masks).unsqueeze(0))
+            x = x.squeeze(0)
+            hxs = hxs.squeeze(0)
         else:
             # x is a (T, N, -1) tensor that has been flatten to (T * N, -1)
             N = hxs.size(0)
@@ -114,18 +116,50 @@ class NNBase(nn.Module):
             x = x.view(T, N, x.size(1))
 
             # Same deal with masks
-            masks = masks.view(T, N, 1)
+            masks = masks.view(T, N)
 
+            # Let's figure out which steps in the sequence have a zero for any agent
+            # We will always assume t=0 has a zero in it as that makes the logic cleaner
+            has_zeros = ((masks[1:] == 0.0) \
+                            .any(dim=-1)
+                            .nonzero()
+                            .squeeze()
+                            .cpu())
+
+            # Otherwise, we will need to process the sequence in smaller chunks
+
+            # +1 to correct the masks[1:]
+            if has_zeros.dim() == 0:
+                # Deal with scalar
+                has_zeros = [has_zeros.item() + 1]
+            else:
+                has_zeros = (has_zeros + 1).numpy().tolist()
+
+            # add t=0 and t=T to the list
+            has_zeros = [0] + has_zeros + [T]
+
+
+            hxs = hxs.unsqueeze(0)
             outputs = []
-            for i in range(T):
-                hx = hxs = self.gru(x[i], hxs * masks[i])
-                outputs.append(hx)
+            for i in range(len(has_zeros) - 1):
+                # We can now process steps that don't have any zeros in masks together!
+                # This is much faster
+                start_idx = has_zeros[i]
+                end_idx = has_zeros[i + 1]
+
+                rnn_scores, hxs = self.gru(
+                    x[start_idx:end_idx],
+                    hxs * masks[start_idx].view(1, -1, 1)
+                )
+
+                outputs.append(rnn_scores)
 
             # assert len(outputs) == T
             # x is a (T, N, -1) tensor
-            x = torch.stack(outputs, dim=0)
+            x = torch.cat(outputs, dim=0)
             # flatten
             x = x.view(T * N, -1)
+            hxs = hxs.unsqueeze(0)
 
         return x, hxs