The currently recurrent policy implementation uses a `GRUCell` and thus a for-loop to evaluate each step in the rollout.  This PR changes the recurrent policy to use a `GRU` (which unfortunately breaks existing pre-trained models) and evaluates the sequence with as few `forward` calls as possible.

In order to reset the hidden state, the `masks` tensor is checked to see if any element has a zero for each step in the rollout (a zero in `masks` corresponds to at a done in the rollout).  If steps `t_1` and `t_2` have at least one done, then steps `t` such that `t_1 <= t < t_2` are run with one call to `forward`.

This results in a massive speed up whenever the number of steps with at least one done is significantly less than the rollout length.  In `PongNoFrameskip-v4` with PPO I see ~1000 FPS while training with this change while I see ~380 FPS with the current implementation.

This does end up being slower if the number of steps with at least one done is very close to the rollout length.  If I artificially set every step to have at least one done, I see ~305 FPS.  If I artificially set every other step to have at least one done, I see ~450 FPS, so the break even is fairly quick.

For `PongNoFrameskip-v4` with A2C, I see ~1860 FPS with this change, and ~1800 FPS with the current implementation.   Due to the longer rollout in PPO, this change is more beneficial for that algorithm, but it doesn't hurt A2C either.

To check if I didn't break anything by doing this, I trained a recurrent policy with PPO in PongNoFrameskip-v4, and the reward curve looks as expected.
![visdom_image](https://user-images.githubusercontent.com/12722529/46248882-ac4a0e00-c3ed-11e8-94da-5e3d0ccb2034.jpg)