@erikwijmans 2018-10-05
Unfortunately I don't have anywhere with MuJoCo installed, so I am not sure how simple that will be for me to run.